get rid of the useless SessionAuth
and rearrange init Auth and App
