--- tests/unit/services/SessionAuthTest.php
@@ -1,23 +1,0 @@
-<?php declare(strict_types=1);
-/**
- * @author Nicolas CARPi <nico-git@deltablot.email>
- * @copyright 2012 Nicolas CARPi
- * @see https://www.elabftw.net Official website
- * @license AGPL-3.0
- * @package elabftw
- */
-
-namespace Elabftw\Services;
-
-use Elabftw\Elabftw\AuthResponse;
-
-class SessionAuthTest extends \PHPUnit\Framework\TestCase
-{
-    public function testTryAuth(): void
-    {
-        $AuthService = new SessionAuth();
-        $authResponse = $AuthService->tryAuth();
-        $this->assertInstanceOf(AuthResponse::class, $authResponse);
-        $this->assertEquals('session', $authResponse->isAuthBy);
-    }
-}
